[JE2-281] Added to the stock archiving and deletion logic a ban on archiving when products are available and a ban on deletion when the stock is included in arrivals or orders

+++ api/src/services/stocks.service.ts
@@ -7,11 +7,15 @@ import { UpdateStockDto } from '../dto/update-stock.dto'
 import { CreateWriteOffDto } from 'src/dto/create-write-off.dto'
 import { StockManipulationService } from './stock-manipulation.service'
 import { WriteOff } from 'src/types'
+import { Arrival, ArrivalDocument } from '../schemas/arrival.schema'
+import { Order, OrderDocument } from '../schemas/order.schema'
 
 @Injectable()
 export class StocksService {
   constructor(
     @InjectModel(Stock.name) private readonly stockModel: Model<StockDocument>,
+    @InjectModel(Arrival.name) private readonly arrivalModel: Model<ArrivalDocument>,
+    @InjectModel(Order.name) private readonly orderModel: Model<OrderDocument>,
     private readonly stockManipulationService: StockManipulationService,
   ) {}
 
@@ -37,6 +41,19 @@ export class StocksService {
     if (stock.isArchived) throw new ForbiddenException('Склад в архиве.')
 
     return stock
+  }
+
+  async isLocked(id: string) {
+    const stock = await this.stockModel.findById(id)
+    if (!stock) throw new NotFoundException('Склад не найден')
+    const arrivals = await this.arrivalModel.find({
+      stock: stock._id,
+    })
+    if (arrivals.length) return true
+    const orders = await this.orderModel.find({
+      stock: stock._id,
+    })
+    return !!orders.length
   }
 
   async getArchivedById(id: string) {
@@ -89,12 +106,20 @@ export class StocksService {
   }
 
   async archive(id: string) {
-    const stock = await this.stockModel.findByIdAndUpdate(id, { isArchived: true })
+    const stock = await this.stockModel.findById(id).exec()
 
     if (!stock) throw new NotFoundException('Склад не найден.')
 
     if (stock.isArchived) throw new ForbiddenException('Склад уже в архиве.')
 
+    const hasActiveProducts = stock.products.some(p => p.amount > 0)
+
+    if (hasActiveProducts) {
+      throw new ForbiddenException('На складе ещё есть товары. Архивация невозможна.')
+    }
+
+    stock.isArchived = true
+    await stock.save()
     return { message: 'Склад перемещен в архив.' }
   }
 
@@ -112,8 +137,20 @@ export class StocksService {
   }
 
   async delete(id: string) {
-    const stock = await this.stockModel.findByIdAndDelete(id)
+    const stock = await this.stockModel.findById(id).exec()
     if (!stock) throw new NotFoundException('Склад не найден.')
+
+    const hasActiveProducts = stock.products.some(p => p.amount > 0)
+    if (hasActiveProducts) {
+      throw new ForbiddenException('На складе есть товары. Удаление невозможно.')
+    }
+
+    const isLocked = await this.isLocked(id)
+    if (isLocked) {
+      throw new ForbiddenException('Склад участвует в поставках или заказах. Удаление невозможно.')
+    }
+
+    await this.stockModel.findByIdAndDelete(id)
     return { message: 'Склад успешно удален.' }
   }
 }
